+++ sky/backends/cloud_vm_ray_backend.py
@@ -1533,7 +1533,7 @@ class CloudVmRayBackend(backends.Backend):
                 # TODO(suquark): once we have sky on PYPI, we should directly
                 # install sky from PYPI.
                 local_wheel_path = wheel_utils.build_sky_wheel()
-            backoff = backend_utils.Backoff(_RETRY_UNTIL_UP_INIT_GAP_SECONDS)
+            backoff = common_utils.Backoff(_RETRY_UNTIL_UP_INIT_GAP_SECONDS)
             attempt_cnt = 1
             while True:
                 # RetryingVmProvisioner will retry within a cloud's regions
